+++ src/widgets/smart-sort/defaultTemplates.ts
@@ -1,5 +1,0 @@
-export default {
-  text: '',
-  button: ({ isSmartSorted }) =>
-    isSmartSorted ? 'See all results' : 'See relevant results',
-};